+++ lib/internal/util_test.ts
@@ -3,7 +3,6 @@ import {
   assertNotEquals,
   assertStrictEquals,
 } from "https://deno.land/std@0.84.0/testing/asserts.ts";
-import { stripColor } from "https://deno.land/std@0.84.0/fmt/colors.ts";
 import * as util from "./util.ts";
 
 Deno.test("kComb", () => {